Issue number: Obtained from: Submitted by: Reviewed by:
Adding methods on RuntimeEndpointInfo to support proxy object for
endpoint implementor object. If proxy object is set, the method invocation
is done on the proxy object. Otherwise, the method invocation is done
on the actual endpoint implementor object. endpoint implementor object is
still used to create dynamic model as usual.