bug fix: 6337665
All the attachments in the message (mtom and wsdl mime) are avaialble to the client and endpoint using MessageContext.MESSAGE_ATTACHMENTS property. Handlers on the way will see the wsdl mime attachments. JAXWS will disable xop encoding before invoking handlers, basically mtom optimization is not applied to the message when invoking handlers. After handler chain is invoked the XOP is turned back to its previous state.