Issue number: Obtained from: Submitted by: Reviewed by:
Part of bug fix : 6356554

The servlet delegate passes GET requets to endpoints to support RESTful
webservices. If there is a query or extra path info, it is sent it to the
endpoint. But if the query contains a wsdl publishing pattern like wsdl, xsd=,
wsdl=, it is handled by WSDLPublisher.