Fixes a NullPointerException in RuntimeWSDLParser caused by previous optimization change for bug JAX_WS-1001 (this reslted in test failures in jaxws-ri-unit-2.2.6-tomcat)
That change created a scenario where RuntimeWSDLParser.resolveWSDL could potentially return a null value for the parser. I have now added a check to make sure that if the parser is null, we will fall back to using the Source to create it (this used to be the behavior before the change that caused the NPE).