barrida review 777
In order to change the RI as *little* as possible:

- in metro-cs: I made a temporary copy of com/sun/xml/ws/api/message/MessageContextFactory and had DISIMessageContextFactory extend it (rather than the RI impl)
   I did that so the return types of the creation methods are com.oracle....

However, the base class (in the RI): com.oracle.webservices.api.message.MessageContextFactory is incorrectly using org.jvnet.ws.EnvelopeStyle (instead of the com.oracle one).  I initially tried to change this, but too many thing broke.  So instead

- I added : createContext(Source m, com.oracle.webservices.api.EnvelopeStyle.Style envelopeStyle);

The other changes reflect that added method and also:

- SOAPVersion needs to handle both versions of EnvelopeStyle until org.jvnet.ws goes away

- Made the "codec" field in Packet temporarily "public" so I can use it in the DISI copy of MessageContextFactory implementation.